Refactor DrawEngine
One can now add new shape to draw using DrawEngine.addShape(IDrawable).
The IDrawable must implement draw() method, which will be called at
every frame. Thanks to this, drawing aim and inventory was extracted
to external classes.

+++ DigitalSimulator/src/pl/jbujak/simulator/input/MouseButtonProcessor.java
@@ -2,7 +2,8 @@ package pl.jbujak.simulator.input;
 
 import org.lwjgl.glfw.GLFWMouseButtonCallback;
 
-import pl.jbujak.simulator.environment.IPlayer;
+import pl.jbujak.simulator.Simulation;
+import pl.jbujak.simulator.player.IPlayer;
 import static org.lwjgl.glfw.GLFW.*;
 
 public class MouseButtonProcessor extends GLFWMouseButtonCallback{
@@ -15,6 +16,9 @@ public class MouseButtonProcessor extends GLFWMouseButtonCallback{
 	@Override
 	public void invoke(long windowHandle, int button, int action, int mods) {
 		
+		if(Simulation.isPaused()) {return;}
+		//Events below will not be processed if simulation is paused.
+
 		if(action == GLFW_PRESS) {
 			if(button == GLFW_MOUSE_BUTTON_1) {
 				controlledPlayer.destroyBlock();
